🚑️ fix: channel types are wrong typed
