+++ mod.ts
@@ -67,6 +67,10 @@ export {
   GuildIntegration
 } from './src/structures/guild.ts'
 export { CategoryChannel } from './src/structures/guildCategoryChannel.ts'
+export {
+  GuildForumChannel,
+  GuildForumTag
+} from './src/structures/guildForumChannel.ts'
 export { NewsChannel } from './src/structures/guildNewsChannel.ts'
 export { VoiceChannel } from './src/structures/guildVoiceChannel.ts'
 export { Invite } from './src/structures/invite.ts'
