+++ src/managers/channelThreads.ts
@@ -6,23 +6,21 @@ import type {
   ThreadChannel,
   ThreadMember
 } from '../structures/threadChannel.ts'
-import type {
-  CreateThreadOptions,
-  GuildTextChannel
-} from '../structures/guildTextChannel.ts'
+import type { CreateThreadOptions } from '../structures/guildTextChannel.ts'
 import type { BaseManager, Message } from '../../mod.ts'
+import type { GuildThreadAvailableChannel } from '../structures/guildThreadAvailableChannel.ts'
 
 export class ChannelThreadsManager extends BaseChildManager<
   ThreadChannelPayload,
   ThreadChannel
 > {
-  channel: GuildTextChannel
+  channel: GuildThreadAvailableChannel
   declare parent: BaseManager<ThreadChannelPayload, ThreadChannel>
 
   constructor(
     client: Client,
     parent: ThreadsManager,
-    channel: GuildTextChannel
+    channel: GuildThreadAvailableChannel
   ) {
     super(
       client,
