+++ src/structures/channel.ts
@@ -43,6 +43,8 @@ import type { VoiceChannel } from '../structures/guildVoiceChannel.ts'
 import type { StageVoiceChannel } from '../structures/guildVoiceStageChannel.ts'
 import type { TextChannel } from '../structures/textChannel.ts'
 import type { ThreadChannel } from '../structures/threadChannel.ts'
+import { CreateInviteOptions } from '../managers/invites.ts'
+import { Invite } from './invite.ts'
 
 export class Channel extends SnowflakeBase {
   type!: ChannelTypes
@@ -440,5 +442,10 @@ export class GuildChannel extends Channel {
   /** Edit position of the channel */
   async setPosition(position: number): Promise<GuildChannels> {
     return await this.edit({ position })
+  }
+
+  /** Create an Invite for this Channel */
+  async createInvite(options?: CreateInviteOptions): Promise<Invite> {
+    return this.guild.invites.create(this.id, options)
   }
 }
