+++ src/structures/guildForumChannel.ts
@@ -1,23 +1,15 @@
+import { Client } from '../client/client.ts'
 import {
-  CategoryChannel,
-  Client,
-  Emoji,
-  Guild,
-  GuildChannel,
   GuildForumChannelPayload,
   GuildForumSortOrderTypes,
   GuildForumTagPayload,
-  Invite,
-  Message,
   ModifyGuildForumChannelOption,
-  ModifyGuildForumChannelPayload,
-  ThreadChannel,
-  ThreadMember
-} from '../../mod.ts'
-import { ChannelThreadsManager } from '../managers/channelThreads.ts'
-import { CreateInviteOptions } from '../managers/invites.ts'
+  ModifyGuildForumChannelPayload
+} from '../types/channel.ts'
 import { CHANNEL } from '../types/endpoint.ts'
-import { CreateThreadOptions } from './guildTextChannel.ts'
+import { Emoji } from './emoji.ts'
+import { Guild } from './guild.ts'
+import { GuildThreadAvailableChannel } from './guildThreadAvailableChannel.ts'
 
 export class GuildForumTag {
   id!: string
@@ -39,12 +31,7 @@ export class GuildForumTag {
   }
 }
 
-export class GuildForumChannel extends GuildChannel {
-  slowmode!: number
-  defaultThreadSlowmode!: number
-  defaultAutoArchiveDuration!: number
-  threads: ChannelThreadsManager
-  topic?: string
+export class GuildForumChannel extends GuildThreadAvailableChannel {
   availableTags!: GuildForumTag[]
   defaultReactionEmoji!: Emoji
   defaultSortOrder!: GuildForumSortOrderTypes
@@ -52,21 +39,10 @@ export class GuildForumChannel extends GuildChannel {
   constructor(client: Client, data: GuildForumChannelPayload, guild: Guild) {
     super(client, data, guild)
     this.readFromData(data)
-    this.threads = new ChannelThreadsManager(
-      this.client,
-      this.guild.threads,
-      this
-    )
   }
 
   readFromData(data: GuildForumChannelPayload): void {
     super.readFromData(data)
-    this.topic = data.topic ?? this.topic
-    this.defaultThreadSlowmode =
-      data.default_thread_rate_limit_per_user ?? this.defaultThreadSlowmode
-    this.defaultAutoArchiveDuration =
-      data.default_auto_archive_duration ?? this.defaultAutoArchiveDuration
-    this.slowmode = data.rate_limit_per_user ?? this.slowmode
     this.availableTags =
       data.available_tags?.map((tag) => new GuildForumTag(tag)) ??
       this.availableTags
@@ -123,167 +99,5 @@ export class GuildForumChannel extends GuildChannel {
     const resp = await this.client.rest.patch(CHANNEL(this.id), body)
 
     return new GuildForumChannel(this.client, resp, this.guild)
-  }
-
-  /** Create an Invite for this Channel */
-  async createInvite(options?: CreateInviteOptions): Promise<Invite> {
-    return this.guild.invites.create(this.id, options)
-  }
-
-  /** Edit topic of the channel */
-  async setTopic(topic: string): Promise<GuildForumChannel> {
-    return await this.edit({ topic })
-  }
-
-  /** Edit category of the channel */
-  async setCategory(
-    category: CategoryChannel | string
-  ): Promise<GuildForumChannel> {
-    return await this.edit({
-      parentID: typeof category === 'object' ? category.id : category
-    })
-  }
-
-  /** Edit Slowmode of the channel */
-  async setSlowmode(slowmode?: number | null): Promise<GuildForumChannel> {
-    return await this.edit({ slowmode: slowmode ?? null })
-  }
-
-  /** Edit Default Slowmode of the threads in the channel */
-  async setDefaultThreadSlowmode(
-    slowmode?: number | null
-  ): Promise<GuildForumChannel> {
-    return await this.edit({ defaultThreadSlowmode: slowmode ?? null })
-  }
-
-  /** Edit Default Auto Archive Duration of threads */
-  async setDefaultAutoArchiveDuration(
-    slowmode?: number | null
-  ): Promise<GuildForumChannel> {
-    return await this.edit({ defaultAutoArchiveDuration: slowmode ?? null })
-  }
-
-  async startThread(
-    options: CreateThreadOptions,
-    message: Message | string
-  ): Promise<ThreadChannel> {
-    const payload = await this.client.rest.endpoints.startPublicThread(
-      this.id,
-      typeof message === 'string' ? message : message.id,
-      { name: options.name, auto_archive_duration: options.autoArchiveDuration }
-    )
-    await this.client.channels.set(payload.id, payload)
-    return (await this.client.channels.get<ThreadChannel>(payload.id))!
-  }
-
-  async startPrivateThread(
-    options: CreateThreadOptions
-  ): Promise<ThreadChannel> {
-    const payload = await this.client.rest.endpoints.startPrivateThread(
-      this.id,
-      { name: options.name, auto_archive_duration: options.autoArchiveDuration }
-    )
-    await this.client.channels.set(payload.id, payload)
-    return (await this.client.channels.get<ThreadChannel>(payload.id))!
-  }
-
-  async fetchArchivedThreads(
-    type: 'public' | 'private' = 'public',
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    const data =
-      type === 'public'
-        ? await this.client.rest.endpoints.getPublicArchivedThreads(
-            this.id,
-            params
-          )
-        : await this.client.rest.endpoints.getPrivateArchivedThreads(
-            this.id,
-            params
-          )
-
-    const threads: ThreadChannel[] = []
-    const members: ThreadMember[] = []
-
-    for (const d of data.threads) {
-      await this.threads.set(d.id, d)
-      threads.push((await this.threads.get(d.id))!)
-    }
-
-    for (const d of data.members) {
-      const thread =
-        threads.find((e) => e.id === d.id) ?? (await this.threads.get(d.id))
-      if (thread !== undefined) {
-        await thread.members.set(d.user_id, d)
-        members.push((await thread.members.get(d.user_id))!)
-      }
-    }
-
-    return {
-      threads,
-      members,
-      hasMore: data.has_more
-    }
-  }
-
-  async fetchPublicArchivedThreads(
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    return await this.fetchArchivedThreads('public', params)
-  }
-
-  async fetchPrivateArchivedThreads(
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    return await this.fetchArchivedThreads('private', params)
-  }
-
-  async fetchJoinedPrivateArchivedThreads(
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    const data =
-      await this.client.rest.endpoints.getJoinedPrivateArchivedThreads(
-        this.id,
-        params
-      )
-
-    const threads: ThreadChannel[] = []
-    const members: ThreadMember[] = []
-
-    for (const d of data.threads) {
-      await this.threads.set(d.id, d)
-      threads.push((await this.threads.get(d.id))!)
-    }
-
-    for (const d of data.members) {
-      const thread =
-        threads.find((e) => e.id === d.id) ?? (await this.threads.get(d.id))
-      if (thread !== undefined) {
-        await thread.members.set(d.user_id, d)
-        members.push((await thread.members.get(d.user_id))!)
-      }
-    }
-
-    return {
-      threads,
-      members,
-      hasMore: data.has_more
-    }
   }
 }
