+++ src/structures/guildNewsChannel.ts
@@ -1,3 +1,8 @@
+import { Mixin } from '../../deps.ts'
 import { GuildTextBasedChannel } from './guildTextChannel.ts'
+import { GuildThreadAvailableChannel } from './guildThreadAvailableChannel.ts'
 
-export class NewsChannel extends GuildTextBasedChannel {}
+export class NewsChannel extends Mixin(
+  GuildTextBasedChannel,
+  GuildThreadAvailableChannel
+) {}
