+++ src/structures/guildTextChannel.ts
@@ -4,20 +4,13 @@ import { GuildChannel } from './channel.ts'
 import type { Client } from '../client/mod.ts'
 import type {
   GuildTextBasedChannelPayload,
-  GuildTextChannelPayload,
   ModifyGuildTextBasedChannelOption,
-  ModifyGuildTextBasedChannelPayload,
-  ModifyGuildTextChannelOption,
-  ModifyGuildTextChannelPayload
+  ModifyGuildTextBasedChannelPayload
 } from '../types/channel.ts'
 import type { Guild } from './guild.ts'
 import { CHANNEL } from '../types/endpoint.ts'
 import type { Message } from './message.ts'
-import type { CreateInviteOptions } from '../managers/invites.ts'
-import type { Invite } from './invite.ts'
-import type { CategoryChannel } from './guildCategoryChannel.ts'
-import type { ThreadChannel, ThreadMember } from './threadChannel.ts'
-import { ChannelThreadsManager } from '../managers/channelThreads.ts'
+import { GuildThreadAvailableChannel } from './guildThreadAvailableChannel.ts'
 
 export interface CreateThreadOptions {
   /** 2-100 character channel name */
@@ -28,8 +21,6 @@ export interface CreateThreadOptions {
 
 /** Represents a Text Channel but in a Guild */
 export class GuildTextBasedChannel extends Mixin(TextChannel, GuildChannel) {
-  topic?: string
-
   constructor(
     client: Client,
     data: GuildTextBasedChannelPayload,
@@ -41,7 +32,6 @@ export class GuildTextBasedChannel extends Mixin(TextChannel, GuildChannel) {
 
   readFromData(data: GuildTextBasedChannelPayload): void {
     super.readFromData(data)
-    this.topic = data.topic ?? this.topic
   }
 
   /** Edit the Guild Text Channel */
@@ -53,9 +43,7 @@ export class GuildTextBasedChannel extends Mixin(TextChannel, GuildChannel) {
       position: options?.position,
       permission_overwrites: options?.permissionOverwrites,
       parent_id: options?.parentID,
-      nsfw: options?.nsfw,
-      topic: options?.topic
-      // rate_limit_per_user: options?.slowmode
+      nsfw: options?.nsfw
     }
 
     const resp = await this.client.rest.patch(CHANNEL(this.id), body)
@@ -98,213 +86,10 @@ export class GuildTextBasedChannel extends Mixin(TextChannel, GuildChannel) {
 
     return this
   }
-
-  /** Create an Invite for this Channel */
-  async createInvite(options?: CreateInviteOptions): Promise<Invite> {
-    return this.guild.invites.create(this.id, options)
-  }
-
-  /** Edit topic of the channel */
-  async setTopic(topic: string): Promise<GuildTextBasedChannel> {
-    return await this.edit({ topic })
-  }
-
-  /** Edit category of the channel */
-  async setCategory(
-    category: CategoryChannel | string
-  ): Promise<GuildTextBasedChannel> {
-    return await this.edit({
-      parentID: typeof category === 'object' ? category.id : category
-    })
-  }
 }
 
-export class GuildTextChannel extends GuildTextBasedChannel {
-  slowmode!: number
-  defaultThreadSlowmode!: number
-  defaultAutoArchiveDuration!: number
-  threads: ChannelThreadsManager
-
-  constructor(client: Client, data: GuildTextChannelPayload, guild: Guild) {
-    super(client, data, guild)
-    this.readFromData(data)
-    this.threads = new ChannelThreadsManager(
-      this.client,
-      this.guild.threads,
-      this
-    )
-  }
-
-  readFromData(data: GuildTextChannelPayload): void {
-    super.readFromData(data)
-    this.slowmode = data.rate_limit_per_user ?? this.slowmode
-    this.defaultThreadSlowmode =
-      data.default_thread_rate_limit_per_user ?? this.defaultThreadSlowmode
-    this.defaultAutoArchiveDuration =
-      data.default_auto_archive_duration ?? this.defaultAutoArchiveDuration
-  }
-
-  /** Edit the Guild Text Channel */
-  async edit(
-    options?: ModifyGuildTextChannelOption
-  ): Promise<GuildTextChannel> {
-    const body: ModifyGuildTextChannelPayload = {
-      name: options?.name,
-      position: options?.position,
-      permission_overwrites: options?.permissionOverwrites,
-      parent_id: options?.parentID,
-      nsfw: options?.nsfw,
-      topic: options?.topic,
-      rate_limit_per_user: options?.slowmode,
-      default_auto_archive_duration: options?.defaultAutoArchiveDuration,
-      default_thread_rate_limit_per_user: options?.defaultThreadSlowmode
-    }
-
-    const resp = await this.client.rest.patch(CHANNEL(this.id), body)
-
-    return new GuildTextChannel(this.client, resp, this.guild)
-  }
-
-  /** Edit Slowmode of the channel */
-  async setSlowmode(slowmode?: number | null): Promise<GuildTextChannel> {
-    return await this.edit({ slowmode: slowmode ?? null })
-  }
-
-  /** Edit Default Slowmode of the threads in the channel */
-  async setDefaultThreadSlowmode(
-    slowmode?: number | null
-  ): Promise<GuildTextChannel> {
-    return await this.edit({ defaultThreadSlowmode: slowmode ?? null })
-  }
-
-  /** Edit Default Auto Archive Duration of threads */
-  async setDefaultAutoArchiveDuration(
-    slowmode?: number | null
-  ): Promise<GuildTextChannel> {
-    return await this.edit({ defaultAutoArchiveDuration: slowmode ?? null })
-  }
-
-  async startThread(
-    options: CreateThreadOptions,
-    message: Message | string
-  ): Promise<ThreadChannel> {
-    const payload = await this.client.rest.endpoints.startPublicThread(
-      this.id,
-      typeof message === 'string' ? message : message.id,
-      { name: options.name, auto_archive_duration: options.autoArchiveDuration }
-    )
-    await this.client.channels.set(payload.id, payload)
-    return (await this.client.channels.get<ThreadChannel>(payload.id))!
-  }
-
-  async startPrivateThread(
-    options: CreateThreadOptions
-  ): Promise<ThreadChannel> {
-    const payload = await this.client.rest.endpoints.startPrivateThread(
-      this.id,
-      { name: options.name, auto_archive_duration: options.autoArchiveDuration }
-    )
-    await this.client.channels.set(payload.id, payload)
-    return (await this.client.channels.get<ThreadChannel>(payload.id))!
-  }
-
-  async fetchArchivedThreads(
-    type: 'public' | 'private' = 'public',
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    const data =
-      type === 'public'
-        ? await this.client.rest.endpoints.getPublicArchivedThreads(
-            this.id,
-            params
-          )
-        : await this.client.rest.endpoints.getPrivateArchivedThreads(
-            this.id,
-            params
-          )
-
-    const threads: ThreadChannel[] = []
-    const members: ThreadMember[] = []
-
-    for (const d of data.threads) {
-      await this.threads.set(d.id, d)
-      threads.push((await this.threads.get(d.id))!)
-    }
-
-    for (const d of data.members) {
-      const thread =
-        threads.find((e) => e.id === d.id) ?? (await this.threads.get(d.id))
-      if (thread !== undefined) {
-        await thread.members.set(d.user_id, d)
-        members.push((await thread.members.get(d.user_id))!)
-      }
-    }
-
-    return {
-      threads,
-      members,
-      hasMore: data.has_more
-    }
-  }
-
-  async fetchPublicArchivedThreads(
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    return await this.fetchArchivedThreads('public', params)
-  }
-
-  async fetchPrivateArchivedThreads(
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    return await this.fetchArchivedThreads('private', params)
-  }
-
-  async fetchJoinedPrivateArchivedThreads(
-    params: { before?: string; limit?: number } = {}
-  ): Promise<{
-    threads: ThreadChannel[]
-    members: ThreadMember[]
-    hasMore: boolean
-  }> {
-    const data =
-      await this.client.rest.endpoints.getJoinedPrivateArchivedThreads(
-        this.id,
-        params
-      )
-
-    const threads: ThreadChannel[] = []
-    const members: ThreadMember[] = []
-
-    for (const d of data.threads) {
-      await this.threads.set(d.id, d)
-      threads.push((await this.threads.get(d.id))!)
-    }
-
-    for (const d of data.members) {
-      const thread =
-        threads.find((e) => e.id === d.id) ?? (await this.threads.get(d.id))
-      if (thread !== undefined) {
-        await thread.members.set(d.user_id, d)
-        members.push((await thread.members.get(d.user_id))!)
-      }
-    }
-
-    return {
-      threads,
-      members,
-      hasMore: data.has_more
-    }
-  }
-}
+// Still exist for API compatibility
+export class GuildTextChannel extends Mixin(
+  GuildTextBasedChannel,
+  GuildThreadAvailableChannel
+) {}
